Added route to Reservations Details page and moved ReservedProductDetails
