+++ code/src/components/reserved-product-details/ReservedProductDetails.scss
