+++ code/src/components/reserved-product-details/ReservedProductDetails.types.ts
