+++ code/src/components/reserved-product-details/messages/en.json
@@ -4,5 +4,6 @@
     "reservedProductDetails.productQuantity.label": "Quantity:",
     "reservedProductDetails.productPrice.label": "Price:",
     "reservedProductDetails.productPriceWithDiscount.label": "Price with Discount:",
-    "reservedProductDetails.productDiscount.label": "Discount percentage:"
+    "reservedProductDetails.productDiscount.label": "Discount percentage:",
+    "reservationsPage.reservationsNotFound": "Reservations details page not found"
 }
