+++ code/src/components/reserved-product-details/messages/index.ts
@@ -1,0 +1,4 @@
+import en from "@/components/reserved-product-details/messages/en.json";
+import uk from "@/components/reserved-product-details/messages/uk.json";
+
+export default { en, uk };
