+++ code/src/messages.ts
@@ -29,7 +29,7 @@ import myAddressesMessages from "@/containers/user-account/my-addresses/messages
 import myReservationsMessages from "@/containers/user-account/my-reservations/messages";
 import imageSearchModalMessages from "@/containers/modals/image-search-modal/messages";
 import reservationsTableMessages from "@/containers/tables/reservations-table/messages"
-import reservationsDetailsMessages from "@/pages/reservations-details/messages";
+import reservationsDetailsMessages from "@/components/reserved-product-details/messages";
 
 import dropDownMessages from "@/components/app-dropdown/messages";
 import bestSellerContainerMessages from "@/components/bestseller-card/messages";
