--- code/src/pages/reservations-details/messages/index.ts
@@ -1,4 +1,0 @@
-import en from "@/pages/reservations-details/messages/en.json";
-import uk from "@/pages/reservations-details/messages/uk.json";
-
-export default { en, uk };
